lq-realtime-service-main/pyproject.toml
[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=app --cov-report=term-missing"

# [build-system]
# requires = ["setuptools>=65.0", "wheel"]
# build-backend = "setuptools.build_meta"

[project]
name = "lq-realtime-service"
version = "1.0.0"
description = "LingoQuesto Real-time WebSocket & Webhook Service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "websockets>=12.0",
    "aioredis>=2.0.1",
    "redis>=5.0.1",
    "PyJWT>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "orjson>=3.9.12",
    "psutil>=7.1.3",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "anyio>=4.0.0",
  "ruff>=0.6.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.11.0",
  "ruff>=0.1.0",
  "mypy>=1.5.0",
]
